Move account transaction list component to components directory and wire it up to router
