+++ src/entry/client/routes.jsx
@@ -1,7 +1,9 @@
 import React from 'react';
 import {mount} from 'react-mounter';
 import Layout from '../../imports/ui/components/layout.jsx';
-import AccountTransactionsList from '../../imports/ui/components/account-transactions-list.jsx';
+import AccountTransactionsList from '../../imports/ui/containers/account-transactions-list';
+
+FlowRouter.wait();
 
 FlowRouter.route("/", {
   action() {
